✅ Test: 배포 환경에서의 테스트 추가
