+++ .gitignore
@@ -274,3 +274,5 @@ node_modules/
 /playwright-report/
 /blob-report/
 /playwright/.cache/
+storageState.local.json
+storageState.staging.json
