+++ src/app/workspace/(workspace-modal-fallback)/terms/page.tsx
@@ -1,20 +1,16 @@
 import { Suspense } from 'react';
 
-import TermsModalPage from '../../@workspaceModal/(.)terms/page';
+import TermsModalPage from '../../[workspaceId]/@innerWorkspaceIdModal/(.)terms/page';
 import MainPage from '../../page';
 
 // 새로고침이나 직접 접근 시에도 TermsModalPage를 재활용
-const TermsStandalonePage = async ({
-  params,
-}: Readonly<{
-  params: Promise<{ workspaceId?: string }>;
-}>) => {
+const TermsStandalonePage = () => {
   return (
     <>
       <Suspense fallback={<div>로딩중...</div>}>
         <TermsModalPage />
       </Suspense>
-      <MainPage params={params} />
+      <MainPage />
     </>
   );
 };
