+++ src/app/workspace/[workspaceId]/(inner-workspaceId-modal-fallback)/search/default.tsx
@@ -1,0 +1,5 @@
+const Default = () => {
+  return null;
+};
+
+export default Default;
