+++ src/app/workspace/[workspaceId]/(inner-workspaceId-modal-fallback)/search/page.tsx
@@ -1,0 +1,18 @@
+import { Suspense } from 'react';
+
+import SearchModalPage from '@/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)search/page';
+
+import MainWithWorkspacePage from '../../page';
+
+const DownloadStandalonePage = () => {
+  return (
+    <>
+      <Suspense fallback={<div>로딩중...</div>}>
+        <SearchModalPage />
+      </Suspense>
+      <MainWithWorkspacePage />
+    </>
+  );
+};
+
+export default DownloadStandalonePage;
