+++ src/app/workspace/[workspaceId]/(inner-workspaceId-modal-fallback)/settings/change-password/page.tsx
@@ -1,0 +1,19 @@
+import { Suspense } from 'react';
+
+import ChangePasswordModalPage from '@/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)settings/(.)change-password/page';
+
+const ChangePasswordStandalonePage = ({
+  params,
+}: {
+  params: Promise<{ workspaceId?: string }>;
+}) => {
+  return (
+    <>
+      <Suspense fallback={<div>로딩중...</div>}>
+        <ChangePasswordModalPage />
+      </Suspense>
+    </>
+  );
+};
+
+export default ChangePasswordStandalonePage;
