+++ src/app/workspace/[workspaceId]/(inner-workspaceId-modal-fallback)/settings/page.tsx
@@ -1,0 +1,15 @@
+import { Suspense } from 'react';
+
+import SettingModalPage from '@/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)settings/page';
+
+const SettingsStandalonePage = ({ params }: { params: Promise<{ workspaceId?: string }> }) => {
+  return (
+    <>
+      <Suspense fallback={<div>로딩중...</div>}>
+        <SettingModalPage />
+      </Suspense>
+    </>
+  );
+};
+
+export default SettingsStandalonePage;
