+++ src/app/workspace/[workspaceId]/(inner-workspaceId-modal-fallback)/terms/page.tsx
@@ -1,0 +1,18 @@
+import { Suspense } from 'react';
+
+import TermsModalPage from '../../@innerWorkspaceIdModal/(.)terms/page';
+import MainWithWorkspacePage from '../../page';
+
+// 새로고침이나 직접 접근 시에도 TermsModalPage를 재활용
+const TermsStandalonePage = () => {
+  return (
+    <>
+      <Suspense fallback={<div>로딩중...</div>}>
+        <TermsModalPage />
+      </Suspense>
+      <MainWithWorkspacePage />
+    </>
+  );
+};
+
+export default TermsStandalonePage;
