--- src/app/workspace/[workspaceId]/(main)/(with-workspace-modal-fallback)/search/page.tsx
@@ -1,17 +1,0 @@
-import { Suspense } from 'react';
-
-import SearchModalPage from '../../@withWorkspaceModal/(.)search/page';
-import MainWithWorkspacePage from '../../page';
-
-const DownloadStandalonePage = () => {
-  return (
-    <>
-      <Suspense fallback={<div>로딩중...</div>}>
-        <SearchModalPage />
-      </Suspense>
-      <MainWithWorkspacePage />
-    </>
-  );
-};
-
-export default DownloadStandalonePage;
