--- src/app/workspace/[workspaceId]/(main)/(with-workspace-modal-fallback)/settings/change-password/page.tsx
@@ -1,19 +1,0 @@
-import { Suspense } from 'react';
-
-import ChangePasswordModalPage from '../../../@withWorkspaceModal/(.)settings/(.)change-password/page';
-
-const ChangePasswordStandalonePage = ({
-  params,
-}: {
-  params: Promise<{ workspaceId?: string }>;
-}) => {
-  return (
-    <>
-      <Suspense fallback={<div>로딩중...</div>}>
-        <ChangePasswordModalPage />
-      </Suspense>
-    </>
-  );
-};
-
-export default ChangePasswordStandalonePage;
