--- src/app/workspace/[workspaceId]/(main)/(with-workspace-modal-fallback)/settings/page.tsx
@@ -1,15 +1,0 @@
-import { Suspense } from 'react';
-
-import SettingModalPage from '../../@withWorkspaceModal/(.)settings/page';
-
-const SettingsStandalonePage = ({ params }: { params: Promise<{ workspaceId?: string }> }) => {
-  return (
-    <>
-      <Suspense fallback={<div>로딩중...</div>}>
-        <SettingModalPage />
-      </Suspense>
-    </>
-  );
-};
-
-export default SettingsStandalonePage;
