--- src/app/workspace/[workspaceId]/(main)/@withWorkspaceModal/default.tsx
@@ -1,5 +1,0 @@
-const Default = () => {
-  return null;
-};
-
-export default Default;
