+++ src/app/workspace/[workspaceId]/@innerWorkspaceIdModal/(.)settings/(.)change-password/default.tsx
@@ -1,0 +1,5 @@
+const Default = () => {
+  return null;
+};
+
+export default Default;
