+++ src/app/workspace/[workspaceId]/@innerWorkspaceIdModal/default.tsx
@@ -1,0 +1,5 @@
+const Default = () => {
+  return null;
+};
+
+export default Default;
