+++ src/app/workspace/[workspaceId]/team-mood-tracker/survey/[moodTrackerHashedId]/download/page.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+import TeamMoodTrackerDetailPage from '@features/team-mood-tracker/components/page-wrapper/TeamMoodTrackerDetailPage/TeamMoodTrackerDetailPage.client';
+
+import DownloadModalPage from '../../../@teamMoodTrackerGeneralModal/(.)survey/[moodTrackerHashedId]/download/page';
+
+const downlaodFallbackPage = () => {
+  return (
+    <div>
+      <DownloadModalPage />
+      <TeamMoodTrackerDetailPage />
+    </div>
+  );
+};
+
+export default downlaodFallbackPage;
