+++ src/app/workspace/layout.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from 'next';
 
-import FooterLayout from '@common/components/layouts/FooterLayout.server';
-
 export const metadata: Metadata = {
   title: '메인 페이지 - HaRu',
   description: '하루의 메인페이지입니다.',
